requirement.txt
# Django and REST framework
Django>=5.1.4
djangorestframework>=3.15.2
django-cors-headers>=4.6.0

# API Documentation
drf-yasg>=1.21.7

# Task Queue and Message Broker
celery>=5.4.0
redis>=5.2.1

# Utilities
python-dotenv>=1.0.1
pytz>=2024.2
python-dateutil>=2.9.0

# Other Dependencies
PyYAML>=6.0.2
uritemplate>=4.1.1
inflection>=0.5.1
six>=1.17.0
sqlparse>=0.5.3
typing-extensions>=4.12.2
asgiref>=3.8.1